refactor: 컴포넌트 생성 경오 `src/component/ui`에서 `src/component`로 변경
