+++ packages/design-system/src/components/button/Button.tsx
@@ -1,5 +1,5 @@
 import { ComponentProps } from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../lib/utils';
 
 interface ButtonProps extends ComponentProps<'button'> {
   size?: 'small' | 'large';
