+++ packages/design-system/src/index.ts
@@ -1,6 +1,6 @@
 import '@lococo/tailwind-config';
 
-export * from './components/ui/index';
+export * from './components/index';
 
 export * from './lib/utils';
